Update from IceFog

+++ api/src/services/users/destroy-service.ts
@@ -1,4 +1,4 @@
-import db, { Role, User, UserGroupMembership } from "@/models"
+import db, { Dataset, Role, User, UserGroupMembership } from "@/models"
 
 import BaseService from "@/services/base-service"
 
@@ -11,6 +11,16 @@ export class DestroyService extends BaseService {
   }
 
   async perform(): Promise<void> {
+    const datasetsWithOwner = await Dataset.findAll({
+      where: { ownerId: this.user.id },
+    })
+    if (datasetsWithOwner.length > 0) {
+      const datasetNames = datasetsWithOwner.map((dataset) => dataset.name).join(", ")
+      throw new Error(
+        `Cannot delete user as they are the owner of the following datasets: ${datasetNames}.`
+      )
+    }
+
     return db.transaction(async () => {
       await UserGroupMembership.destroy({ where: { userId: this.user.id } })
       await Role.destroy({ where: { userId: this.user.id } })
